Refactor AtomicQueue
